Draft of a ComparisonTable class
Quickly implemented a (non-functional) ComparisonTable class and an
instance of it. So the structure would be:

* the repository `stacks-table` contains the `ComparisonTable` class
* the repository `morphism-properties-preservation-table` overloads the
  `ComparisonTable` class
* wherever you want to output a table you load the overloaded PHP file
  and call `OverloadedTable->outputTable()` (I foresee some problems
  with multiple definitions in this construction...)

It doesn't work (I haven't even checked the syntax), but it shows that a
generic approach to the problem isn't too hard. The only thing left to
do before there is a working version is insert the database handling
code, and we're off to go!

I now realise that maybe it would also be possible to have the base
repository as a submodule. So maybe a configuration value pointing to
*some* copy of the base repository would be the best solution, if one
wishes to use multiple ComparisonTables in `stacks-website`.